feat: Refactor DailyReporter and environment variable checks; add new structure for daily report generation
