+++ src/daily_report/main.py
@@ -1,0 +1,5 @@
+from .daily_reporter import DailyReporter
+
+if __name__ == "__main__":
+    reporter = DailyReporter()
+    reporter.run()
