--- src/main.py
@@ -1,4 +1,0 @@
-from daily_reporter import DailyReporter
-
-if __name__ == "__main__":
-    reporter = DailyReporter()
